Move Tabs controllers to organization directories
