+++ test/screens/Home/Tabs/Workspaces/workspacesController.test.js
@@ -5,7 +5,7 @@ const {
   getTimeAgoString,
   updateTimestamps,
   setFavorite,
-} = require("../../../../../src/screens/Home/Tabs/Workspaces/workspacesController.js");
+} = require("../../../../../src/screens/Home/Tabs/Workspaces/controllers/workspacesController.js");
 
 class TextEncoderMock {
   encode(str) {
